Put processed data for specific peekbank dataset on OSF

Put processed data for specific peekbank dataset on OSF

ds.put_processed_data(token, dataset_name, path = ".", osf_address = "pr6wu")

Arguments

token

personal access tokens for uploading to OSF

dataset_name

Specific dataset name occurring in the file hierarchy of the relevant OSF repo.

path

Where the data live on your own machine.

osf_address

pr6wu for peekbank.