Get AOI data

Get AOI data

get_aoi_data(dataset_id = NULL, lab_dataset_id = NULL, age = NULL,
  connection = NULL)

Arguments

dataset_id
An integer vector of one or more dataset ids

lab_dataset_id
A character vector of one or more lab dataset ids

age
A numeric vector of a single age or a min age and max age (inclusive), in months

connection
A connection to the peekbank database

Value

A `tbl` of AOI data, filtered down by supplied arguments. If `connection` is supplied, the result remains a remote query, otherwise it is retrieved into a local tibble.

Examples

get_aoi_data(lab_dataset_id = "pomper_saffran_2016")

#> # A tibble: 226,809 x 23
#>    aoi_data_id aoi       t subject_id trial_id lab_trial_id distractor_image
#>          <int> <chr> <int>      <int>    <int> <chr>        <chr>           
#>  1           0 targ…  -990          0        0 SwitchingCu… crib            
#>  2          64 targ…  -990          4        0 SwitchingCu… crib            
#>  3         335 dist…  -990         21        0 SwitchingCu… crib            
#>  4         750 targ…  -990         47        0 SwitchingCu… crib            
#>  5         845 dist…  -990         53        0 SwitchingCu… crib            
#>  6         957 targ…  -957          0        0 SwitchingCu… crib            
#>  7        1021 targ…  -957          4        0 SwitchingCu… crib            
#>  8        1292 dist…  -957         21        0 SwitchingCu… crib            
#>  9        1707 targ…  -957         47        0 SwitchingCu… crib            
#> 10        1802 dist…  -957         53        0 SwitchingCu… crib            
#> # … with 226,799 more rows, and 16 more variables: distractor_label <chr>,
#> #   full_phrase <int>, point_of_disambiguation <int>, target_image <chr>,
#> #   target_label <chr>, target_side <chr>, aoi_region <int>, dataset_id <int>,
#> #   monitor_size_x <int>, monitor_size_y <int>, sample_rate <int>,
#> #   tracker <int>, lab_dataset_id <chr>, age <int>, sex <chr>,
#> #   lab_subject_id <chr>