Get XY timepoints

Get XY timepoints

get_xy_timepoints(
  dataset_id = NULL,
  dataset_name = NULL,
  age = NULL,
  connection = NULL
)

Arguments

dataset_id
An integer vector of one or more dataset ids

dataset_name
A character vector of one or more dataset names

age
A numeric vector of a single age or a min age and max age (inclusive), in months

connection
A connection to the peekbank database

Value

A `tbl` of XY timepoints data, filtered down by supplied arguments. If `connection` is supplied, the result remains a remote query, otherwise it is retrieved into a local tibble.

Examples

if (FALSE) {
get_xy_timepoints(dataset_name = "pomper_saffran_2016")
}